Separate inlining and profiling phases
